Moving HITS work to nlp.py and test_nlp.py
